requirements.txt
PyQt6
requests
websocket-client
Pillow
screeninfo
pynput
numpy
opencv-python
websockets
dbus-next
mss
PyGObject
pyobjc
